+++ blacksheep/server/application.py
@@ -30,7 +30,7 @@ from guardpost.common import AuthenticatedRequirement
 from itsdangerous import Serializer
 from rodi import ContainerProtocol
 
-from blacksheep.baseapp import BaseApplication, handle_not_found
+from blacksheep.baseapp import BaseApplication, default_fallback, handle_not_found
 from blacksheep.common import extend
 from blacksheep.common.files.asyncfs import FilesHandler
 from blacksheep.contents import ASGIContent
@@ -639,9 +639,26 @@ class Application(BaseApplication):
         configured_handlers.clear()
 
     def _normalize_fallback_route(self):
+        """
+        Automatically configures the NotFound exception handler to use the user-defined
+        fallback route when no explicit NotFound handler is defined.
+        This provides intuitive behavior where custom fallback routes handle 404s.
+
+        The purpose of this function is to improve the user experience, as a
+        user-defined fallback route is automatically used to handle NotFound
+        exceptions.
+        """
         fallback = self.router.fallback
 
-        if fallback is not None and self._has_default_not_found_handler():
+        if (
+            fallback is not None
+            and fallback.handler is not default_fallback
+            and self._has_default_not_found_handler()
+        ):
+            self.logger.debug(
+                "The user-defined fallback route is configured to handle NotFound "
+                "exceptions."
+            )
 
             async def fallback_handler(app, request, exc) -> Response:
                 return await fallback.handler(request)  # type: ignore